more tweaks
git-svn-id: http://svn.research.cc.gatech.edu/kaos/atl/trunk@3292 8ba6a8b9-845b-0410-94f0-974be2f5df47

+++ attr.c
@@ -1079,7 +1079,8 @@ recursive_encode(attr_list l, AttrBuffer b, attr_value_type t)
 	    int attr_count = l->l.list.iattrs->int_attr_count;
 	    void *buffer_end;
 	    if (attr_count == 0) return;
-	    buffer_end = add_to_tmp_buffer(b, (int) attr_count * sizeof(int_attr));
+	    buffer_end = add_to_tmp_buffer(b, (unsigned)(attr_count * 
+					   sizeof(int_attr)));
 	    memcpy(buffer_end, &l->l.list.iattrs->iattr[0], 
 		   attr_count * sizeof(int_attr));
 	    ((int_attr_p) b->tmp_buffer)->int_attr_count += attr_count;
@@ -1152,7 +1153,7 @@ encode_attr_for_xmit(attr_list l, AttrBuffer b, int *length)
 	    return l->l.list.iattrs;
 	}
     }
-    add_to_tmp_buffer(b, sizeof(struct int_attr_struct));
+    add_to_tmp_buffer(b, (unsigned)sizeof(struct int_attr_struct));
     ((int_attr_p) b->tmp_buffer)->int_attr_count = 0;
     ((int_attr_p) b->tmp_buffer)->other_attr_count = 0;
     /* not the simple case */
